Work...

+++ ptm/adventure_managment/spells/implementations/create_travel.py
@@ -2,16 +2,17 @@
 Module contains spell and recipe for create travel. Travel is python module
 containing rules for scrapping.
 '''
-from os.path import join
+from os.path import join, isfile
+from shutil import copy2
 
-from . import Spell
-from .recipe import Recipe, Ingredient
-from .types import LOCAL_SPELL
+from .. import Spell
+from ..recipe import Recipe, Ingredient
+from ..types import LOCAL_SPELL
 
-from ...ptm_settings import TRAVEL_TEMPLATE
+from ....ptm_settings import TRAVEL_TEMPLATE
 
 
-class CreateTravel:
+class CreateTravel(Spell):
     name = 'create_travel'
     description = '''Create travel. Travel is a python module
                           containing scrapping rules.'''
@@ -23,5 +24,9 @@ class CreateTravel:
 
     def execute(self):
         land_path = self.recipe.land or ''
-        travel_path = join(join(land, self.rucksack.TRAVELS_PATH), '{}.py'.format(self.recipe.name))
-        copy2(TRAVEL_TEMPLATE, travel_path)
+        travel_path = join(join(self.rucksack.TRAVELS_PATH, land_path),
+                           '{}.py'.format(self.recipe.name))
+        if not isfile(travel_path):
+            copy2(TRAVEL_TEMPLATE, travel_path)
+        else:
+            print('Package {} is exist!'.format(travel_path))
